requirements.txt
streamlit
openai
keras-nlp
keras
kagglehub